+++ CHANGELOG.rst
@@ -1,10 +1,10 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package ros2_test_framework
+Changelog for package rtest
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 0.1.0 (2018-05-07)
 ------------------
-* Initial release of ros2_test_framework package.
+* Initial release of rtest package.
 * Supports running tests in ROS 2 using the Google Test framework.
 * Provides mocks for ROS 2 publishers and subscribers.
 * Provides mocks for ROS 2 services and actions.